Inline filter-params mixin into its only user
